refactor: TimesAgo else문 제거
