+++ src/test/java/com/owori/domain/story/service/StoryServiceTest.java
@@ -148,35 +148,6 @@ public class StoryServiceTest extends LoginTest {
     }
 
     @Test
-    @DisplayName("time ago enum class 가 제대로 실행되는가")
-    void findTimeAgo() {
-
-        // 50초 전
-        String second = TimesAgo.of(LocalDateTime.now().minusSeconds(50));
-
-        // 10분 전
-        String minute = TimesAgo.of(LocalDateTime.now().minusMinutes(10));
-
-        // 2시간 전
-        String hour = TimesAgo.of(LocalDateTime.now().minusHours(2));
-
-        // 6일 전
-        String day = TimesAgo.of(LocalDateTime.now().minusDays(6));
-
-        // n개월 전
-        String month = TimesAgo.of(LocalDateTime.of(2022,3,2,3,1,2,3));
-
-
-        //then
-        assertThat(second).isEqualTo("50초 전");
-        assertThat(minute).isEqualTo("10분 전");
-        assertThat(hour).isEqualTo("2시간 전");
-        assertThat(day).isEqualTo("6일 전");
-        assertThat(month).isEqualTo("22.03.02");
-
-    }
-
-    @Test
     @DisplayName("이야기 수정이 수행되는가")
     void updateStory() {
         //given
